@@ -1,5 +1,5 @@
 class InvitationsController < ApplicationController
-  before_filter :authenticate_user!
+  before_action :authenticate_user!
   
   def new
     if I18n.locale == :es
