@@ -1,5 +1,6 @@
 class UsersController < ApplicationController
- before_filter :authenticate_user!
+ before_action :authenticate_user_from_token!
+ before_action :authenticate_user!
 
   def show
     @user = User.find(params[:id])
