@@ -1,22 +1,24 @@
-require File.expand_path('../boot', __FILE__)
+require_relative 'boot'
 
+require "rails"
 # Pick the frameworks you want:
-# require "active_record/railtie"
-require "action_controller/railtie"
+require "active_model/railtie"
+require "active_job/railtie"
+#require "active_record/railtie"
 require "action_mailer/railtie"
-require "active_resource/railtie"
+require "action_view/railtie"
+#require "action_cable/engine"
 require "sprockets/railtie"
-# require "rails/test_unit/railtie"
+require "rails/test_unit/railtie"
 
-if defined?(Bundler)
-  # If you precompile assets before deploying to production, use this line
-  Bundler.require *Rails.groups(:assets => %w(development test staging))
-  # If you want your assets lazily compiled in production, use this line
-  # Bundler.require(:default, :assets, Rails.env)
-end
+# Require the gems listed in Gemfile, including any gems
+# you've limited to :test, :development, or :production.
+Bundler.require(*Rails.groups)
 
 module Eclickr
   class Application < Rails::Application
+    # Initialize configuration defaults for originally generated Rails version.
+    config.load_defaults 5.1
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
@@ -41,7 +43,8 @@ module Eclickr
     config.i18n.load_path += Dir[Rails.root.join('config', 'locales', '**', '*.{rb,yml}')]
     #config.i18n.default_locale = :de
 
-    config.i18n.enforce_available_locales = false
+    # Whitelist locales available for the application
+    I18n.available_locales = [:en, :de]
 
     # Configure the default encoding used in templates for Ruby 1.9.
     config.encoding = "utf-8"
