@@ -1,8 +1,9 @@
-# Load the rails application
-require File.expand_path('../application', __FILE__)
+# Load the Rails application.
+require_relative 'application'
 
 # Initialize the rails application
-Eclickr::Application.initialize!
+Rails.application.initialize!
+#Eclickr::Application.initialize!
 
 # https://github.com/logentries/le_ruby
 if ENV["LOGENTRIES_TOKEN"] && ENV["LOGENTRIES_TOKEN"] != "" && defined?(Le)
