@@ -1,5 +1,7 @@
-Eclickr::Application.configure do
+Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb
+  
+  config.eager_load = false
 
   # In the development environment your application's code is reloaded on
   # every request.  This slows down response time but is perfect for development
@@ -29,6 +31,8 @@ Eclickr::Application.configure do
   config.assets.debug = true
   
   config.action_mailer.default_url_options = { :host => 'localhost:3000' }
+  
+  config.public_file_server.enabled = true
 
   # Raise exception on mass assignment protection for Active Record models
   # config.active_record.mass_assignment_sanitizer = :strict
