@@ -1,5 +1,7 @@
-Eclickr::Application.configure do
+Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb
+  
+  config.eager_load = true
 
   # Code is not reloaded between requests
   config.cache_classes = true
@@ -66,6 +68,8 @@ Eclickr::Application.configure do
   config.lograge.custom_options = lambda do |event|
     {ip: event.payload[:ip]}
   end
+  
+  config.active_record.dump_schema_after_migration = false
 
 end
 
