@@ -1,5 +1,7 @@
-Eclickr::Application.configure do
+Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb
+  
+  config.eager_load = true
 
   # Code is not reloaded between requests
   config.cache_classes = true
