@@ -1,5 +1,7 @@
 Eclickr::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb
+  
+  config.eager_load = false
 
   # The test environment is used exclusively to run your application's
   # test suite.  You never need to work with it otherwise.  Remember that
@@ -7,9 +9,13 @@ Eclickr::Application.configure do
   # and recreated between test runs.  Don't rely on the data there!
   config.cache_classes = true
 
-  # Configure static asset server for tests with Cache-Control for performance
-  config.serve_static_assets = true
-  config.static_cache_control = "public, max-age=3600"
+  # Configure public file server for tests with Cache-Control for performance.
+
+  config.public_file_server.enabled = true
+  config.public_file_server.headers = {
+    'Cache-Control' => "public, max-age=#{1.hour.seconds.to_i}"
+  }
+ 
 
   # Log error messages when you accidentally call methods on nil
   config.whiny_nils = true
