@@ -1,4 +1,6 @@
 # http://stackoverflow.com/a/8713889/238931
 Warden.test_mode!
-World Warden::Test::Helpers
+
+include Warden::Test::Helpers
+
 After { Warden.test_reset! }
